move mcp service into its own package
