+++ internal/service/mcp/mcp.go
@@ -1,4 +1,4 @@
-package service
+package mcp
 
 import (
 	"fmt"
