+++ internal/service/mcp/util_test.go
@@ -1,4 +1,4 @@
-package service
+package mcp
 
 import (
 	"testing"
